compiling 2 version of the programm and adjust some settings for v1.1

+++ seed_generator.cpp
@@ -15,7 +15,7 @@ GridSeed generate_random_seed(int width, int height, int percent_alive) {
 
 	// Fill the grid
 	std::srand(std::time(nullptr));
-	for (int i = 0; i < grid_seed.grid.size(); i += 4) {
+	for (size_t i = 0; i < grid_seed.grid.size(); i += 4) {
 		if((std::rand() % 100 + 1) <= percent_alive) {
 			grid_seed.grid[i] = 1;
 		}
